Web: Files: refactoring Tile, fix drag, moved EditingWraperComponent to sub-components
