+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/FilesTile/BadgesFileTile.js
